+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.7.0
 -----
 - Bug #37: Fixed `conquer\oauth2\TokenAuth` behavior when multiple Authorization headers are present
+- Bug #32: `conquer\oauth2\BaseModel` caches request
 - Enh #24: Client Credentials grant type support
 - Enh: Added `scopes` setting to `conquer\oauth2\TokenAuth`
 - Enh: Added migrations namespace